Add dummy test
I would have liked to add a more meaningful test, i.e. using
DropwizardAppRule or ResourceTestRule, however I bumped into
annoying bugs and not being able to work around them within
the time I could spend on this.
Given this is just a test repository which will never be used
on production, this should be fine.